+++ test/testdata/misc.py
@@ -1,7 +1,8 @@
 import abc
-from functools import lru_cache
+from functools import lru_cache, singledispatchmethod
 from typing import Generic
 from typing import TypeVar
+from typing import Union
 
 from pdoc._compat import cached_property
 
@@ -385,6 +386,32 @@ def another_decorated_function(arg: str) -> str:
     pass
 
 
+class SingleDispatchMethodExample:
+    @singledispatchmethod
+    def fancymethod(self, str_or_int: Union[str, int]):
+        """A fancy method which is capable of handling either `str` or `int`.
+
+        :param str_or_int: string or integer to handle
+        """
+        raise NotImplementedError(f"{type(str_or_int)=} not implemented!")
+
+    @fancymethod.register
+    def fancymethod_handle_str(self, str_to_handle: str):
+        """Fancy method handles a string.
+
+        :param str_to_handle: string which will be handled
+        """
+        print(f"{type(str_to_handle)} = '{str_to_handle}")
+
+    @fancymethod.register
+    def _fancymethod_handle_int(self, int_to_handle: int):
+        """Fancy method handles int (not shown in doc).
+
+        :param int_to_handle: int which will be handled
+        """
+        print(f"{type(int_to_handle)} = '{int_to_handle:x}'")
+
+
 __all__ = [
     "Issue226",
     "var_with_default_obj",
@@ -417,4 +444,5 @@ __all__ = [
     "repr_not_syntax_highlightable",
     "ClassDecorator",
     "another_decorated_function",
+    "SingleDispatchMethodExample",
 ]